Use float32 for denoise blending

+++ climg/denoise_stub.go
@@ -5,4 +5,4 @@ package climg
 import "image"
 
 // denoiseImage is a stub used when the nodenoise build tag is set.
-func denoiseImage(img *image.RGBA, sharpness, maxPercent float64) {}
+func denoiseImage(img *image.RGBA, sharpness, maxPercent float32) {}
